rename top-level Redux `vet360` slice to `vapService`
