+++ src/platform/user/profile/vap-svc/components/AddressField/AddressEditModal.jsx
@@ -126,7 +126,7 @@ class AddressEditModal extends React.Component {
 }
 
 const mapStateToProps = state => ({
-  modalData: state.vet360?.modalData,
+  modalData: state.vapService?.modalData,
 });
 
 export default connect(mapStateToProps)(AddressEditModal);
